pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "brunner"
version = "1.0.0"
requires-python = ">=3.10"
dependencies = [
    "beaker-py>=2.0.1",
    "beaker-gantry",
    "rich",
    "pyyaml",
]

[project.scripts]
brunner = "beaker_runner.__main__:main"
brunner-launch = "beaker_runner.launch:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["beaker_runner*"]

[tool.gantry]
